+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.2] - 2022-05-24
+* Update license
+
+## [2.0.1] - 2022-05-16
+**Breaking Change:** Instance data on players inventory items is now passed in as an `object` and returned as a `IDeserializable` 
+* Pre/exp tags have been removed
+* Removed update instance section in inventory UI sample
+* Fixed backwards compatible interface for `Economy.Purchases`
+
 ## [2.0.1-exp.2] - 2022-05-11
 * Add missing samples file